Create Problem 3
Evaluate the overlap integral for two 1s orbitals as a function of interatomic spacing, R. Remember that this is a 3D orbital.
Compare your computational result to the analytic result.

Explore the effect of both the number of guesses and the size of the bounding box as the extent of the orbital goes out to infinity. Evaluation of this integral is often helped by utilizing a weighting function.